Farm project/server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Flask-CORS==4.0.0

# Database drivers
psycopg2-binary==2.9.9  # PostgreSQL
# sqlite3 is included with Python

# Optional: Production server
Gunicorn==21.2.0

# Optional: Security
python-decouple==3.8
